[model] 프로젝트 최종 코드 반영
