+++ Vehicle Detection/CarPlate_utils.py
@@ -1,33 +1,28 @@
 import torch
-from torch.utils.data import Dataset, DataLoader
 from torchvision.datasets.vision import VisionDataset
 import transforms as T
 import glob
-import os
 from PIL import Image
-import numpy as np
 import random
 import os
-import cv2
-import  xml.dom.minidom
 import xml.etree.ElementTree as ET
+
+
+# Car (승용차), Truck (트럭), Bus (버스), Etc vehicle (기타 차량-덤프트럭, 레미콘 등 건설용차량), Bike( 이륜 차) License(번호판)
 
 class ConvertCarPlatetoCOCO(object):
     CLASSES = (
-        "__background__", "Car", "Number Plate",
+        "__background__", "Car", "Truck", "Bus", "Etc vehicle", "Bike", "License",
     )
-    def __call__(self, image, annotation_path):
-        # return image, target
-        # annotation_files =  os.listdir(annotation_path)
-        # f = open(label_path, 'r') # 텍스트파일을 연다
-        # annotation_files = [car0.xml, car1.xml........ ] 이 반환됨
 
+    def __call__(self, image, annotation_path):
         boxes = []
         classes = []
         # for annotation_file in annotation_files:
-            # xml_path = os.path.join(annotation_path, annotation_files)
-        
+        # xml_path = os.path.join(annotation_path, annotation_files)
+
         doc = ET.parse(annotation_path)
+
         root = doc.getroot()
 
         # root = dom.documentElement
@@ -36,44 +31,31 @@ class ConvertCarPlatetoCOCO(object):
         filename = root.findtext("filename")
 
         for object in root.iter("object"):
-            # bndbox = root.getElementsByTagName('bndbox')[0]
-            # xmin = bndbox.getElementsByTagName('xmin')[0]
-            # ymin = bndbox.getElementsByTagName('ymin')[0]
-            # xmax = bndbox.getElementsByTagName('xmax')[0]
-            # ymax = bndbox.getElementsByTagName('ymax')[0]
-
-            # xmin_data = xmin.childNodes[0].data
-            # ymin_data = ymin.childNodes[0].data
-            # xmax_data = xmax.childNodes[0].data
-            # ymax_data = ymax.childNodes[0].data
-
-            xmin = int(float(object.find("bndbox").findtext("xmin")))
-            ymin = int(float(object.find("bndbox").findtext("ymin")))
-            xmax = int(float(object.find("bndbox").findtext("xmax")))
-            ymax = int(float(object.find("bndbox").findtext("ymax")))
+            xmin = float(object.find("bndbox").findtext("xmin"))
+            ymin = float(object.find("bndbox").findtext("ymin"))
+            xmax = float(object.find("bndbox").findtext("xmax"))
+            ymax = float(object.find("bndbox").findtext("ymax"))
             bbox = [xmin, ymin, xmax, ymax]
-        
-            boxes.append(bbox)
-            classes.append(1) # car index label = 1
-                
 
-            
-        boxes =torch.as_tensor(boxes, dtype=torch.float32)
+            boxes.append(bbox)
+            classes.append(self.CLASSES.index(object.findtext("name")))
+
+        boxes = torch.as_tensor(boxes, dtype=torch.float32)
         classes = torch.as_tensor(classes)
 
         target = {}
         target['boxes'] = boxes
         target['labels'] = classes
-        target['name'] = torch.tensor([ord(i) for i in list(filename)], dtype=torch.int8) #convert filename in int8 
+        target['name'] = torch.tensor([ord(i) for i in list(filename)], dtype=torch.int8)  # convert filename in int8
 
         return image, target
 
+
 class CarPlateDetection(VisionDataset):
-    def __init__(self, img_folder, all_images_path, all_labels_path, image_set, transforms = None):
-        print("############################## \n","CarPlateDetection Init")
+    def __init__(self, img_folder, all_images_path, all_labels_path, image_set, transforms=None):
+        print("############################## \n", "CarPlateDetection Init")
         # print(img_folder, image_set, transforms)
 
-        
         self.all_images_path = all_images_path
         self.all_labels_path = all_labels_path
 
@@ -84,13 +66,13 @@ class CarPlateDetection(VisionDataset):
         label_path = self.all_labels_path[idx]
 
         img = Image.open(data_path)
-        img = img.convert("RGB") # image BGR -> RGB convert
+        img = img.convert("RGB")  # images BGR -> RGB convert
 
         if self._transforms is not None:
             img, target = self._transforms(img, label_path)
-            
+
         return img, target
-    
+
     def get_height_and_width(self, idx):
         data_path = self.all_images_path[idx]
 
@@ -99,6 +81,7 @@ class CarPlateDetection(VisionDataset):
 
     def __len__(self):
         return len(self.all_images_path)
+
 
 # get CarPlate dataset
 def get_CarPlate(root, image_set, transforms):
@@ -109,46 +92,42 @@ def get_CarPlate(root, image_set, transforms):
     transforms = T.Compose(t)
 
     print("Split CarPlate Data")
-    images_train_path, images_test_path, labels_train_path, labels_test_path = get_CarPlate_image_path_split_list(img_folder=root, test_size=0.2)
+    images_train_path, images_test_path, labels_train_path, labels_test_path = get_CarPlate_image_path_list(
+        img_folder=root)
 
     print("train data count:", len(images_train_path), "/ test data count:", len(images_test_path))
 
     if image_set == "train":
         print("Create Train Dataset Car Plate Detection")
-        dataset = CarPlateDetection(img_folder=root, all_images_path=images_train_path, 
-        all_labels_path=labels_train_path, image_set=image_set, transforms=transforms)
+        dataset = CarPlateDetection(img_folder=root, all_images_path=images_train_path,
+                                    all_labels_path=labels_train_path, image_set=image_set, transforms=transforms)
     else:
         print("Create Test Dataset Car Plate Detection")
-        dataset = CarPlateDetection(img_folder=root, all_images_path=images_test_path, 
-        all_labels_path=labels_test_path, image_set=image_set, transforms=transforms)
+        dataset = CarPlateDetection(img_folder=root, all_images_path=images_test_path,
+                                    all_labels_path=labels_test_path, image_set=image_set, transforms=transforms)
     print("###############################")
 
-    
     return dataset
 
+
 # CarPlate data split
-def get_CarPlate_image_path_split_list(img_folder, test_size= 0.2, seed=0):
-    all_images_path = sorted(glob.glob(os.path.join(img_folder, "../../Vehicle/Car/image", "*")))
-    all_labels_path = sorted(glob.glob(os.path.join(img_folder, "../../Vehicle/Car/annotations", "*")))
+def get_CarPlate_image_path_list(img_folder, seed=0):
+    train_images_path = sorted(glob.glob(os.path.join(img_folder, "train/images", "*")))
+    train_labels_path = sorted(glob.glob(os.path.join(img_folder, "train/annotations", "*")))
+    test_images_path = sorted(glob.glob(os.path.join(img_folder, "test/images", "*")))
+    test_labels_path = sorted(glob.glob(os.path.join(img_folder, "test/annotations", "*")))
 
-    if len(all_images_path) != len(all_labels_path):
-        raise Exception("Split Data Failed")
+    if len(train_images_path) != len(train_labels_path):
+        raise Exception("Train Data Error")
 
-    if test_size < 0 or test_size > 1:
-        raise Exception("Split Test Size 0 ~ 1")
-        
-    total_data = list(zip(all_images_path, all_labels_path))
+    if len(test_images_path) != len(test_labels_path):
+        raise Exception("Test Data Error")
+
+    total_data = list(zip(train_images_path, train_labels_path))
     random.seed(seed)
     random.shuffle(total_data)
 
-    images_path, labels_path = map(list, zip(*total_data))
-
-    divide_index = int(len(all_images_path) * (1-test_size)) # train index / test index
-
-    x_train = images_path[:divide_index]
-    x_test = images_path[divide_index:]
-
-    y_train = labels_path[:divide_index]
-    y_test = labels_path[divide_index:]
+    x_train, y_train = map(list, zip(*total_data))
+    x_test, y_test = list(test_images_path), list(test_labels_path)
 
     return x_train, x_test, y_train, y_test
