+++ Vehicle Detection/engine.py
@@ -1,14 +1,6 @@
-from utils import MetricLogger
+
 import math
 import sys
-import time
-import torch
-
-import torchvision.models.detection.mask_rcnn
-
-from coco_utils import get_coco_api_from_dataset
-from coco_eval import CocoEvaluator
-from voc_eval import _write_voc_results_file, _do_python_eval
 import utils
 
 
@@ -26,206 +18,35 @@ def train_one_epoch(model, optimizer, data_loader, device, epoch, print_freq):
         lr_scheduler = utils.warmup_lr_scheduler(optimizer, warmup_iters, warmup_factor)
 
     for images, targets in metric_logger.log_every(data_loader, print_freq, header):
-        images = list(image.to(device) for image in images)
-        targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
+        try:
+            images = list(image.to(device) for image in images)
+            targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
 
-        loss_dict = model(images, targets)
+            loss_dict = model(images, targets)
 
-        losses = sum(loss for loss in loss_dict.values())
+            losses = sum(loss for loss in loss_dict.values())
 
-        # reduce losses over all GPUs for logging purposes
-        loss_dict_reduced = utils.reduce_dict(loss_dict)
-        losses_reduced = sum(loss for loss in loss_dict_reduced.values())
+            # reduce losses over all GPUs for logging purposes
+            loss_dict_reduced = utils.reduce_dict(loss_dict)
+            losses_reduced = sum(loss for loss in loss_dict_reduced.values())
 
-        loss_value = losses_reduced.item()
+            loss_value = losses_reduced.item()
 
-        if not math.isfinite(loss_value):
-            print("Loss is {}, stopping training".format(loss_value))
-            print(loss_dict_reduced)
-            sys.exit(1)
+            if not math.isfinite(loss_value):
+                print("Loss is {}, stopping training".format(loss_value))
+                print(loss_dict_reduced)
+                sys.exit(1)
 
-        optimizer.zero_grad()
-        losses.backward()
-        optimizer.step()
+            optimizer.zero_grad()
+            losses.backward()
+            optimizer.step()
 
-        if lr_scheduler is not None:
-            lr_scheduler.step()
+            if lr_scheduler is not None:
+                lr_scheduler.step()
 
-        metric_logger.update(loss=losses_reduced, **loss_dict_reduced)
-        metric_logger.update(lr=optimizer.param_groups[0]["lr"])
+            metric_logger.update(loss=losses_reduced, **loss_dict_reduced)
+            metric_logger.update(lr=optimizer.param_groups[0]["lr"])
+        except:
+            pass
 
     return metric_logger
-
-
-def _get_iou_types(model):
-    model_without_ddp = model
-    if isinstance(model, torch.nn.parallel.DistributedDataParallel):
-        model_without_ddp = model.module
-    iou_types = ["bbox"]
-    if isinstance(model_without_ddp, torchvision.models.detection.MaskRCNN):
-        iou_types.append("segm")
-    if isinstance(model_without_ddp, torchvision.models.detection.KeypointRCNN):
-        iou_types.append("keypoints")
-    return iou_types
-
-
-@torch.no_grad()
-def coco_evaluate(model, data_loader, device):
-    n_threads = torch.get_num_threads()
-    # FIXME remove this and make paste_masks_in_image run on the GPU
-    torch.set_num_threads(1)
-    cpu_device = torch.device("cpu")
-    model.eval()
-    metric_logger = utils.MetricLogger(delimiter="  ")
-    header = 'Test:'
-
-    coco = get_coco_api_from_dataset(data_loader.dataset)
-    iou_types = _get_iou_types(model)
-    coco_evaluator = CocoEvaluator(coco, iou_types)
-
-    for images, targets in metric_logger.log_every(data_loader, 100, header):
-        images = list(img.to(device) for img in images)
-
-        if torch.cuda.is_available():
-            torch.cuda.synchronize()
-        model_time = time.time()
-        outputs = model(images)
-
-        outputs = [{k: v.to(cpu_device) for k, v in t.items()} for t in outputs]
-        model_time = time.time() - model_time
-
-        res = {target["image_id"].item(): output for target, output in zip(targets, outputs)}
-        evaluator_time = time.time()
-        coco_evaluator.update(res)
-        evaluator_time = time.time() - evaluator_time
-        metric_logger.update(model_time=model_time, evaluator_time=evaluator_time)
-
-    # gather the stats from all processes
-    metric_logger.synchronize_between_processes()
-    print("Averaged stats:", metric_logger)
-    coco_evaluator.synchronize_between_processes()
-
-    # accumulate predictions from all images
-    coco_evaluator.accumulate()
-    coco_evaluator.summarize()
-    torch.set_num_threads(n_threads)
-    return coco_evaluator
-
-
-@torch.no_grad()
-def voc_evaluate(model, data_loader, device):
-    n_threads = torch.get_num_threads()
-    torch.set_num_threads(1)
-    cpu_device = torch.device("cpu")
-    model.eval()
-    metric_logger = utils.MetricLogger(delimiter="  ")
-    header = 'Test:'
-
-    all_boxes = [[] for i in range(21)]
-    image_index = []
-    for image, targets in metric_logger.log_every(data_loader, 100, header):
-        image = list(img.to(device) for img in image)
-        targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
-
-        torch.cuda.synchronize()
-        model_time = time.time()
-        outputs = model(image)
-
-        name = ''.join([chr(i) for i in targets[0]['name'].tolist()])
-        image_index.append(name)
-
-        outputs = [{k: v.to(cpu_device) for k, v in t.items()} for t in outputs]
-
-        image_boxes = [[] for i in range(21)]
-        for o in outputs:
-            for i in range(o['boxes'].shape[0]):
-                image_boxes[o['labels'][i]].extend([
-                    torch.cat([o['boxes'][i], o['scores'][i].unsqueeze(0)], dim=0)
-                ])
-
-        # makes sure that the all_boxes is filled with empty array when
-        # there are no boxes in image_boxes
-        for i in range(21):
-            if image_boxes[i] != []:
-                all_boxes[i].append([torch.stack(image_boxes[i])])
-            else:
-                all_boxes[i].append([])
-
-        model_time = time.time() - model_time
-
-    metric_logger.synchronize_between_processes()
-
-    all_boxes_gathered = utils.all_gather(all_boxes)
-    image_index_gathered = utils.all_gather(image_index)
-
-    # results from all processes are gathered here
-    if utils.is_main_process():
-        all_boxes = [[] for i in range(21)]
-        for abgs in all_boxes_gathered:
-            for ab, abg in zip(all_boxes, abgs):
-                ab += abg
-        image_index = []
-        for iig in image_index_gathered:
-            image_index += iig
-
-        _write_voc_results_file(all_boxes, image_index, data_loader.dataset.root,
-                                data_loader.dataset._transforms.transforms[0].CLASSES)
-        _do_python_eval(data_loader)
-    torch.set_num_threads(n_threads)
-
-@torch.no_grad()
-def CarPlate_evaluate(model, data_loader, device):
-    n_threads = torch.get_num_threads()
-    torch.set_num_threads(1)
-    cpu_device = torch.device("cpu")
-    model.eval()
-    metric_logger = utils.MetricLogger(delimiter=" ")
-    header ="Test"
-    all_boxes = [[] for i in range(13)]
-    image_index =[]
-
-    for image, targets in metric_logger.log_every(data_loader, 100, header):
-        image = list(img.to(device) for img in image)
-        targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
-
-        torch.cuda.synchronize()
-        model_time = time.time()
-        outputs = model(image)
-
-        name = ''.join([chr(i) for i in targets[0]['name'].tolist()])
-        image_index.append(name)
-
-        outputs = [{k: v.to(cpu_device) for k, v in t.items()} for t in outputs]
-
-        image_boxes = [[] for i in range(13)]
-        for o in outputs:
-            for i in range(o['boxes'].shape[0]):
-                image_boxes[o['labels'][i]].extend([torch.cat([o['boxes'][i], o['scores'][i].unsqueeze(0)], dim=0)])
-
-
-        for i in range(13):
-            if image_boxes[i] != []:
-                all_boxes[i].append([torch.stack(image_boxes[i])])
-            else:
-                all_boxes[i].append([])
-
-        model_time = time.time() - model_time
-
-    metric_logger.synchronize_between_processes()
-
-    all_boxes_gathered = utils.all_gather(all_boxes)
-    image_index_gathered = utils.all_gather(image_index)
-
-    if utils.is_main_process():
-        all_boxes = [[] for i in range(13)]
-        for abgs in all_boxes_gathered:
-            for ab, abg in zip(all_boxes, abgs):
-                ab += abg
-        image_index = []
-        for iig in image_index_gathered:
-            image_index += iig
-
-        _write_voc_results_file(all_boxes, image_index, data_loader.dataset.root,
-                                data_loader.dataset._transforms.transforms[0].CLASSES)
-        _do_python_eval(data_loader)
-    torch.set_num_threads(n_threads)
